Add python3 support for TIMIT data generation
There were two errors here.  The first is that we were opening a
binary audio file without declaring it as binary.  In python3 this
causes the file to be treated as UTF-8, and so we fail due to decoding
errors.  Second, when we convert the byte array to a list of integers,
the peculiar method we use of encoding the bytes as hex and decoding
them as integers is python2-specific and unneeded in python3.

This commit allows the generation of data for TIMIT problems to work
under python3.