Add python3 support for CELEBA generator
We were failing to open image files in binary mode, which in python3
causes the files to be interpreted as UTF-8, causing us to hit
decoding errors when we read from the file.  This commit causes us to
open the image files needed for CELEBA in binary mode, allowing the
CELEBA data to be generated under python3.