Add python3 support for COCO token problems
We were failing to open image files in binary mode, which in python3
causes the files to be interpreted as UTF-8, causing us to hit
decoding errors when we read from the file.  This commit causes us to
open the image files needed for COCO in binary mode, allowing the COCO
data to be generated under python3.